Authorize a user using Google Drive account

Logs a user into their Google account for authorized use of the app. Allows app content to be made conditionally available to 3 classes of user:

authGoogleUI(id)

authGoogle(input, output, session, appURL, authFile)

Arguments

id
An ID string that corresponds with the ID used to call the module's UI function.

input
shiny server input

output
shiny server output

session
shiny server session

appURL
URL to the hosted app.

authFile
Private url to a Google Drive file with access control.

Details

1. Anonymous (non-logged in) users; 2. Logged in users who are not autheticated against a whitelist; 3. Logged in and authenticated users.

This mechanism relies on the file permissions set on a file in Google Drive. Simply create a blank text file (e.g., name_of_your_app.txt) in Google Drive and edit the access control settings to allow only authorized users to download/view this file. You will need the private URL (link) to this file. Assign this URL to a variable in your global.R and pass that as the authFile argument when calling the authGoogle module.

This allows access control to be managed via Google rather than locally on the shiny server.

Login status is stored in session$userData$userLoggedIn(), a reactive value. Authentication status is stored in session$userData$userAuthorized(), a reactive value.

Note

Based on googleAuthR::googleAuth and googleAuthR::googleAuthUI.

Additional requirements

Your global.R file should set the following options: 1. `googleAuthR.scopes.selected`: `c("https://www.googleapis.com/auth/drive.readonly", "https://www.googleapis.com/auth/userinfo.email", "https://www.googleapis.com/auth/userinfo.profile")`. 2. `googleAuthR.webapp.client_id`: your Google app oauth id. 3. `googleAuthR.webapp.client_secret`: your Google app oauth "secret".

Be sure to also set appURL and authFile in global.R.

See the authentication vignette (vignette("authentication", "SpaDES.shiny")).