Histogram Module UI Function

UI function of a shiny module which creates a histogram using barplot based on the data received

Server function of a shiny module which creates a histogram using barplot based on the data received.

histogramUI(id, ...)

histogram(input, output, session, histdata, addAxisParams = NULL,
  verticalBar = NULL, fname = NULL, ...)

Arguments

id
An id string that corresponds with the id used to call the module's UI function.

...
Additional UI parameters passed to link[shiny]{plotOutput}, or additional server arguments passed to barplot.

input
shiny server input object

output
shiny server output object

session
shiny server session object

histdata
Reactive value containing a numeric vector of proportions corresponding to each histogram bin (i.e., the output of hist(..., plot = FALSE) represented as a proportion). Desired subtables can be retrieved using chosenCategories and chosenValues parameters.

addAxisParams
Reactive value with parameters to axis. If NULL (default) then no axis is drawn.

verticalBar
Numeric value at which to add an abline to the histogram.

fname
optional filepath to save .png outputs

Value

None. Invoked for the side-effect of rendering bar plot.