Histogram for Raster Shiny Module

A shiny module creating histogram of data contained in a raster.

Histogram for Raster Shiny Module

histogramForRasterUI(id, title = "", plotParameters, ...)

histogramForRaster(input, output, session, raster, histogramBreaks, scale = 1,
  addAxisParams = NULL, ...)

Arguments

id
An ID string that corresponds with the ID used to call the module server function.

title
Optional title for the histogram. Any shiny tag can be used.

plotParameters
A list of parameters passed to plotOutput.

...
Additional parameters passed to box in histogramForRasterUI, and barplot in histogramForRaster.

input
Shiny server input object

output
Shiny server output object

session
Shiny server session object

raster
Reactive value containing raster

histogramBreaks
Reactive value which is responsible for breaks parameter, as in hist.

scale
Number used for scaling heights of histogram bars. When set to 1 (default) histogram bar height represents the number of raster cells with value from bar interval. If the resolution of raster is known, scale can be used to transform these heights into the ones representing area covered by cells (count). When set to 2 (or, more generally, some number n with no further meaning) this will just increase the height of each histogram bar by 2 (n, respectively). So, in this scenario, each histogram's bar height is just count times 2 (count times n).

addAxisParams
Reactive value with parameters to axis. If NULL (default) then no axis is drawn.

Value

None. Invoked for the side-effect of creating shiny UI.

None. Invoked for the side-effect of rendering histogram plot.