Get installed package dependencies

This is simply a wrapper around tools::package_dependencies

This is simply a wrapper around utils::packageVersion to 1) allow a vector of package names to be passed as an argument; 2) to warn instead of error when a package is not installed; and 3) return versions as character strings.

Get authors from a package's DESCRIPTION

pkgDeps(x, type = NULL)

pkgVers(x, lib.loc = NULL)

pkgAuthors(x)

Arguments

x
An installed package name.

type
One of "Depends", "Imports", "LinkingTo", "Suggests", "Enhances", or "all". Default is NULL, which returns all.

lib.loc
a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages and namespaces are searched before the libraries.

Value

A character vector of package dependencies.

A character vector of package dependencies.

A person vector of package authors.