Tiles updater (shiny module)

Creates a shiny module which uses leafletProxy to update tiles displayed on a leaflet map.

tilesUpdater(input, output, session, proxy, rctUrlTemplateSingleRaster,
  tilesGroup, addTilesParameters = NULL,
  addLayersControlParameters = NULL)

Arguments

input
shiny server input object

output
shiny server output object

session
shiny server session object

proxy
leaflet proxy which manages a connected leaflet map.

rctUrlTemplateSingleRaster
Reactive value which contains the URL template with the tiles (for a single raster) that should be added to the leaflet map as a result.

tilesGroup
Group of rasters/tiles we want to update.

addTilesParameters
Additional parameters of addTiles function. Parameters: proxy/map, urlTemplate and group should be provided directly as tilesUpdater arguments using proxy, urlTemplate and tilesGroup parameter respectively. If passed within addTilesParameters, these arguments will be suppressed by the values passed directly to tilesUpdater.

addLayersControlParameters
Additional parameters passed to addLayersControl.

Value

None. Invoked for the side-effect of sending information to leaflet proxy.

Note

This is a server-only module with no UI component.