Time series of rasters shiny module

A shiny module that displays rasters changing in time on a predefined map, with the user choosing which raster should be displayed by using the slider. A histogram summary for each raster choice is also shown.

timeSeriesofRastersUI(id)

timeSeriesofRasters(input, output, session, rctRasterList, rctUrlTemplate,
  rctPolygonList, rctChosenPoly, defaultPolyName = NULL,
  shpStudyRegionName = NULL, shpStudyRegionLFLT = NULL, colorPalette,
  maxAge, zoom = 5, mapTilesDir = "www/", mapLegend = "",
  mapTitle = "", sliderTitle = "", histTitle = "", nPolygons,
  nRasters, rasterStepSize = 10, uploadOpts = list(auth = NULL, path =
  NULL, user = NULL), rctStudyArea = NULL, thinKeep = 0.05)

Arguments

id
An id string that corresponds with the id used to call the module's UI function.

input
shiny server input object

output
shiny server output object

session
shiny server session object

rctRasterList
A reactive that gives a list of rasters to be displayed.

rctUrlTemplate
The reactive url template for leaflet map tiles

rctPolygonList
Reactive list with sets of polygons to be displayed on a leaflet map. # TODO: describe the format of the list!

rctChosenPoly
The user-selected polygon.

defaultPolyName
Name of the polygon to use as the default for mapping.

shpStudyRegionName
Name of the study area region (from rctRasterList).

shpStudyRegionLFLT
SpatialPolygonDataFrame for the study region.

colorPalette
Colour palette to use.

maxAge
Maximum simulation age.

zoom
Initial leaflet zoom.

mapTilesDir
Directory where pre-generated map tiles are stored (default "www/"; note the trailing slash). Be sure to add this dir as 'tiles' using shiny::addResourcePath in your global.R. E.g., shiny::addResourcePath("tiles", "www/All/FULL/map-tiles") (note we don't need the trailing slash here).

mapLegend
The legend text to add to the leaflet map.

mapTitle
Title to be shown above the map.

sliderTitle
Title to be shown above the slider.

histTitle
Title to be shown above the histogram.

nPolygons
The number of available polygons.

nRasters
The number of available rasters.

rasterStepSize
Size of step in the raster slider.

uploadOpts
A list of options for use with file uploads: auth logical indicating whether user is authorized to upload; path a directory path to use for file uploads; user the current username (used for creating user-specific paths). The default for all options is NULL, which means do not use.

rctStudyArea
A reactive SpatialPolygons* object for the whole study area region.

thinKeep
Proportion of points in polygons to keep when thinning or plotting. See ms_simplify.

Value

Invoked for the side-effect of creating a shiny UI.

Reactive polygon selected by the user with the polygonChooser module. Invoked for the side-effect of creating shiny server and UI components. # TODO: reword