Time-Since-Fire shiny module

A shiny module showing the time-since-fire values from a raster.

Function timeSinceFire creates a shiny module server function which displays rasters changing in time on a predefined map. A slider, on which you can choose which raster should be currently displayed, is also created. Moreover, a histogram summary for each raster choice is shown.

timeSinceFireUI(id, rastersNumber)

timeSinceFire(input, output, session, rasters, polygonsList, shpStudyRegionFull,
  colorTableFile, timeSinceFirePalette, maxAge, leafletZoomInit = 5,
  studyArea = "SMALL", sim = NULL)

Arguments

id
An ID string that corresponds with the ID used to call the module server function.

rastersNumber
How many rasters can we choose from.

input
Shiny server input object.

output
Shiny server output object.

session
Shiny server session object.

rasters
Set of rasters to be displayed.

polygonsList
List with sets of polygons. Each such set can be displayed on a leaflet map.

shpStudyRegionFull
Study area region.

colorTableFile
File that contains color values for tiles.

timeSinceFirePalette
Color palette for time since fire.

maxAge
Maximum simulation age.

leafletZoomInit
Initial leaflet zoom.

studyArea
Size of study area. Options: "FULL", "EXTRALARGE", "LARGE", "MEDIUM", "NWT", "SMALL".

sim
A SpaDES simulation object (simList).

Value

None. Invoked for the side-effect of creating a shiny UI.

None. Invoked for the side-effect of creating a shiny server part.