Slicer Module

One can imagine behaviour of this module in the following way. A tree of height m is created. We begin at the top of the tree with the entire data table. A category (column) from a data table is chosen. This choice is determined by an user using uiSequence argument. Each value of this fixed category determines a subtable of the data table. For each value choice a child node is created. Value choice is passed to the corresponding child node. Therefore every child node implicitly receives a subtable corresponding to the received value choice. Child node also receives an information from user which category should be fixed next. It then performs the same calculations as the top node, but based on the implicitly received subtable. Simultanously, for each node desired UI (e.g. tabs) is created. If no information about next category to fix is provided for a node, it assumes it is a leaf. At the end, at each leaf, a summary function is applied. Note that each leaf (and each node) receives the entire data table and the information about chosen values leading to this leaf (node). Hence summary can be based on implicitly determined subtable, but also based on entire data table or some subtable determined by a subset of chosen values.

slicerUI(id)

slicer(input, output, session, data, categoryValue, uiSequence, serverFunction,
  uiFunction, chosenCategories = NULL, chosenValues = NULL)

Arguments

id
An ID string that corresponds with the ID used to call the module's UI function

input
Shiny server input object.

output
Shiny server output object.

session
Shiny server session object.

data
Reactive value containing data in form of a data.table. This data table is not changed. Its subtables are accessed using chosenCategories and chosenValues arguments. This is helpful, because the end summary function might require information about entire data table.

categoryValue
Each time the data table is sliced (one dimension is cut off), concrete value of the category is set. This argument stores this value.

uiSequence
A data.table of the form data.table(category = list_of_categories, uiType = list_of_ui_actions). Both lists should contain elements of type character. The category column should contain names of the categories which will be subsequently fixed. The uiType column should contain corresponding UI which should be applied for each category choice. Currently there are two possible UI types to perform: "tab" and "box". Type "box": should be used only together with shinydashboard. An example of proper uiSequence is data.table(category = c("Alliance", "Kingdom"), uiType = c("tab", "box")).

serverFunction
A summary module server function. This function should take three arguments: data, chosenCategories and chosenValues. Inside the function there should be a call to shiny module server function. See example section and compare with link[shiny]{callModule}).

uiFunction
A summary module function UI. This function should take one argument: na. Inside the function there should be a call to shiny module UI function. See example section.

chosenCategories
A list with categories names that were already chosen. Default is NULL.

chosenValues
A list with categories values that were already chosen. Default is NULL.

Value

Shiny module UI.

Shiny module server function.