Tiles Updater Module

Creates a shiny module which uses leafletProxy to update tiles displayed on a leaflet map.

tilesUpdater(input, output, session, proxy, urlTemplate, tilesGroup,
  addTilesParameters = NULL, addLayersControlParameters = NULL)

Arguments

input
Shiny server input object.

output
Shiny server output object.

session
Shiny server session object.

proxy
Leaflet proxy which manages a connected leaflet map.

urlTemplate
Reactive value which contains an url with the tiles that should be added to the leaflet map as a result.

tilesGroup
Group of rasters/tiles we want to update

addTilesParameters
Additional parameters of addTiles function. Parameters: proxy/map, urlTemplate and group should be provided directly as tilesUpdater arguments using proxy, urlTemplate and tilesGroup parameter respectively. If passed within addTilesParameters, these arguments will be surpressed by the values passed directly to tilesUpdater.

addLayersControlParameters
Additional parameters passed to addLayersControl.

Value

None. Invoked for the side-effect of sending information to leaflet proxy.